Fix data cycle calculation when CYCLE_WEEKLY
The cycleDay is referred as a "day of month" (1-31) even when we change to
CYCLE_WEEKLY. The calculations are expecting a day of week (1-7).

Without this fix, we can fall into the (inconvenient) result of using a
completely wrong value for the cycle limit (and corresponding data calculation)
or the worse infinite loop that totally breaks data usage.

Change-Id: I41e0012805236b14f5e0d52ec39683fe6a96f3aa